Consolidate index renumbering
